+++ src/libmidi/MidiUtil.h
@@ -69,7 +69,8 @@ enum MidiErrorCode {
    MidiError_InputError,
    MidiError_InvalidInputErrorBehavior,
 
-   MidiError_RequestedTempoFromNonTempoEvent
+   MidiError_RequestedTempoFromNonTempoEvent,
+   MidiError_RequestedKeySignatureFromNonKeySignatureEvent
 };
 
 class MidiError : public std::exception {